temporal/scripts/structured/tabular-classification/xgboost/tree-based/requirements.txt
scikit-learn==1.3.0
pandas==1.3.5
numpy==1.21.6
joblib==1.3.2
mlflow==2.8.0
python-dotenv==1.0.0
xgboost
seaborn
docker==6.1.3

 
